Rework properties
