+++ Packages/UGF.Property/Runtime/Property.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 35e65b54056dba6418c5d6973fc0e419
+guid: b2f749b2cac0ae44a8c1262e2126958e
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
