+++ Packages/UGF.Property/Runtime/PropertyValueAccess.cs
@@ -1,0 +1,9 @@
+﻿namespace UGF.Property.Runtime
+{
+    public enum PropertyValueAccess
+    {
+        All = 0,
+        ReadOnly = 1,
+        WriteOnly = 2
+    }
+}
