+++ Packages/UGF.Property/Runtime/PropertyValueAccess.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b70534358e074529b631c473bae59940
+timeCreated: 1625074581
